+++ components/persistor/HISTORY.md
@@ -1,3 +1,6 @@
+## 5.4.0
+* Reverting __version__ of all the objects getting updated in the transaction.
+* Changing tracking updates the __org__ properties only if the commit succeed.
 ## 5.3.0
 * objects can return the raw insert sqls if consumers want to track them.
 * changing the structure of the returned object for the sqls list in the commit.